# feat: create custom query to generate number of txs